Update www/js/js.js
In this commit I have implemented a cors function. This allows for cross original resource sharing. I have check to see if it works, and it does. I have pushed it to my android phone and it works, markers for NT are plotted on the map.

Side note - I originally just gave the fetch statement the github json raw path and it worked as well. So I am not sure if i need cors, but I think I have demonstrated how to use it so i will leave it there.