Changes to css and index - see comments
Added an ID to image container div as i needed to remove it using js

The css in this commit is to just make everything inside the html element fit when rotating the screen and various changes to the buttons and headings. I did this to make it look more mobile.