Update www/markers_nt.json
I had to update the markers_json file that was in the master branch because this is where the raw file path is pointed to. Interesting to note as i was making changes to the json file within  the branch but wasn't noticing the changes on my application and this was because my json path was pointing to the master branch json.  So any changes i make to my json file i need to make in the master branch.